Added in sequence numbers so the front end user can make sure they are getting requests in order

+++ cmd/api/room.go
@@ -1,7 +1,7 @@
 package main
 
 type Room struct {
-	Sequence int
+	Sequence uint64
 	ID       string
 	Players  map[string]*Client
 	Game     *GameState
